+++ skforecast/stats/autocorrelation/__init__.py
@@ -1,7 +1,0 @@
-from ._autocorrelation import acf, pacf, calculate_lag_autocorrelation
-
-__all__ = [
-    "acf",
-    "pacf",
-    "calculate_lag_autocorrelation",
-]